Create CustomWebView class Add .idea to gitignore
