+++ app/src/main/AndroidManifest.xml
@@ -22,6 +22,7 @@
         <activity
             android:name="com.jsc.smartpanel.SettingsActivity"
             android:label="@string/title_activity_settings"
+            android:screenOrientation="landscape"
             android:theme="@style/PreferenceScreenStyle">
 
         </activity>
@@ -29,6 +30,7 @@
             android:name="com.jsc.smartpanel.FullscreenActivity"
             android:configChanges="orientation|keyboardHidden|screenSize"
             android:label="@string/app_name"
+            android:screenOrientation="landscape"
             android:theme="@style/FullscreenTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
