+++ app/src/main/java/com/jsc/smartpanel/CustomWebView.java
@@ -1,0 +1,106 @@
+/*
+ * Copyright (c) 2020.
+ * RF Controls
+ * Design and Programming by Alex Dovby
+ */
+
+package com.jsc.smartpanel;
+
+import android.annotation.SuppressLint;
+import android.content.Context;
+import android.os.Build;
+import android.support.annotation.RequiresApi;
+import android.util.AttributeSet;
+import android.webkit.WebResourceRequest;
+import android.webkit.WebSettings;
+import android.webkit.WebView;
+import android.webkit.WebViewClient;
+
+class CustomWebView extends WebView {
+    public CustomWebView(Context context) {
+        super(context);
+        init();
+    }
+
+    public CustomWebView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+        init();
+    }
+
+    public CustomWebView(Context context, AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
+        init();
+    }
+
+//    public CustomWebView(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+//        super(context, attrs, defStyleAttr, defStyleRes);
+//      init();
+//    }
+
+    public CustomWebView(Context context, AttributeSet attrs, int defStyleAttr, boolean privateBrowsing) {
+        super(context, attrs, defStyleAttr, privateBrowsing);
+        init();
+    }
+
+    @SuppressLint("SetJavaScriptEnabled")
+    private void init() {
+        // set WebView Style background and scrollbar ----
+        setVerticalScrollBarEnabled(false);
+        setHorizontalScrollBarEnabled(false);
+
+        //webView.setScrollBarStyle(WebView.SCROLLBARS_OUTSIDE_OVERLAY);
+        //webView.setScrollbarFadingEnabled(false);
+        // set default black color ------------
+        setBackgroundColor(0);
+
+        // web settings --------------------------------------
+        WebSettings webSettings = getSettings();
+
+        // ??? ---------------------------------------
+        webSettings.setLoadsImagesAutomatically(true);
+        webSettings.setAllowFileAccess(true);
+        webSettings.setAppCacheEnabled(true);
+        // ??? ---------------------------------------
+
+        // включаем поддержку JavaScript
+        getSettings().setJavaScriptEnabled(true);
+        // определим экземпляр MyWebViewClient.
+        // Он может находиться в любом месте после инициализации объекта WebView
+        setWebViewClient(new MyWebViewClient());
+//        webView.setWebChromeClient(new WebChromeClient() {
+//            @Override
+//            public void onProgressChanged(WebView view, int newProgress) {
+//                //change your progress bar
+//                Log.w(TAG, "onProgressChanged : " + newProgress + "%");
+//                if (newProgress > 50) {
+//                    //progressBar.setVisibility(View.GONE);
+//                    webView.setVisibility(View.VISIBLE);
+//                }
+//            }
+//        });
+
+        webSettings.setSupportZoom(false);
+        webSettings.setBuiltInZoomControls(false);
+        webSettings.setLoadsImagesAutomatically(true);
+        webSettings.setAllowFileAccess(true);
+
+        // set cash app ---------------------------------------
+        webSettings.setAppCacheEnabled(true);
+        // set wiewport scale ---------------
+        // webSettings.setUseWideViewPort(isWideViewPortRequired());
+        webSettings.setUseWideViewPort(true);
+        webSettings.setLoadWithOverviewMode(true);
+
+    }
+
+    // ----------------------------------------
+    private class MyWebViewClient extends WebViewClient {
+        //        @TargetApi(19)
+        @RequiresApi(api = Build.VERSION_CODES.N)
+        @Override
+        public boolean shouldOverrideUrlLoading(WebView view, WebResourceRequest request) {
+            view.loadUrl(request.getUrl().toString());
+            return true;
+        }
+    }
+}
