+++ app/src/main/res/layout/activity_fullscreen.xml
@@ -21,13 +21,10 @@
         android:layout_height="match_parent"
         android:orientation="vertical">
 
-        <WebView
+        <com.jsc.smartpanel.CustomWebView
             android:id="@+id/web_view"
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-        </WebView>
-
+            android:layout_height="match_parent" />
     </LinearLayout>
 
     <LinearLayout
