+++ app/version.properties
@@ -1,5 +1,5 @@
-#Sat Jan 25 21:55:10 EET 2020
+#Sun Jan 26 09:41:08 PST 2020
 MAJOR=0
 MINOR=9
 VERSION_CODE=152
-PATCH=29
+PATCH=36
